requirements.txt
langchain_openai
smolagents
google-api-python-client
python-dotenv